requirements.txt
keras~=2.15.0
tensorflow~=2.15.0
tensorflow-hub~=0.15.0
bert-tensorflow
tf-models-official
scikit-learn~=1.3.2
pandas~=2.1.4
numpy~=1.26.3
pydot~=2.0.0
graphviz~=0.20.1
matplotlib~=3.8.2
datasets~=2.16.0
jupyter
cassandra-driver~=3.29.0
flask~=3.0.0
apiflask~=2.1.0
psutil~=5.9.7
python-dotenv~=1.0.0